test-requirements.txt
factory-boy
pytest-selenium
pytest-flask
pytest-cov
pytest
tox
